extract subclasses: QbmTempConfig, QuickBackupMultiConfig extract backup logic to RestoreTask create ClientRestoreDelegate to avoid environment issue rename: QuickBackupMultiCommandManager.java -> QuickBackupMultiCommand fix: Recursive call in QbmManager#restore
