+++ src/main/java/dev/skydynamic/quickbackupmulti/i18n/LangSuggestionProvider.java
@@ -1,9 +1,10 @@
-package dev.skydynamic.quickbackupmulti.utils;
+package dev.skydynamic.quickbackupmulti.i18n;
 
 import com.mojang.brigadier.context.CommandContext;
 import com.mojang.brigadier.suggestion.SuggestionProvider;
 import com.mojang.brigadier.suggestion.Suggestions;
 import com.mojang.brigadier.suggestion.SuggestionsBuilder;
+import dev.skydynamic.quickbackupmulti.i18n.Translate;
 import net.minecraft.server.command.ServerCommandSource;
 
 import java.util.concurrent.CompletableFuture;
